requirements.txt
requests
boto3
fastapi
uvicorn
google-generativeai
python-dotenv
